requirements.txt
Django==1.8.3
djangorestframework==3.1.3
drf-extensions==0.2.7
ipython==3.2.1
psycopg2==2.6.1
wheel==0.24.0
